updated-latest

+++ app/Http/Controllers/Site/Pages/ContactController.php
@@ -2,29 +2,36 @@
 
 namespace App\Http\Controllers\Site\Pages;
 
-use App\Exceptions\NotAuthrizedException;
-use App\Exceptions\NotFoundException;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ContactRequest;
 use App\Models\Contact;
 use App\Models\Provider;
 use Auth;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ContactController extends Controller
 {
     public function index()
     {
         try {
-            $userId= Auth::id();
+            $userId = Auth::id();
             $verifiedContacts = Contact::whereUserId($userId)->whereIsVerified(1)->paginate(3, ['*'], 'verified');
             $unVerifiedContacts = Contact::whereUserId($userId)->whereIsVerified(0)->paginate(3, ['*'], 'unverified');
             $providers = Provider::get();
-            return view('site.pages.contactInfo',compact('providers','unVerifiedContacts','verifiedContacts'));
+            return view('site.pages.contactInfo', compact('providers', 'unVerifiedContacts', 'verifiedContacts'));
         } catch (\Exception $e) {
             return $e;
         }
     }
+
+    public function addedContact(){
+        try{
+            return view('site.pages.addedContact');
+        }catch(\Exception $e){
+            return $e;
+        }
+    }
+
     public function store(ContactRequest $request)
     {
         try {
@@ -33,7 +40,7 @@ class ContactController extends Controller
                 'provider_id' => $request->providerId,
                 'user_id' => Auth::id()
             ]);
-            if($saveContact){
+            if ($saveContact) {
                 return response()->json([
                     'status' => true,
                     'msg' => 'added successfully',
@@ -49,38 +56,77 @@ class ContactController extends Controller
     public function show($id)
     {
         try {
-            $contact = $this->showContact($id);
-            return response()->json($contact);
-        } catch (NotFoundException $e) {
+            $userId = Auth::id();
+            $contact = Contact::whereUserId($userId)->whereId($id)->first();
+            if(!$contact){
+                return response()->json([
+                        'status' => false,
+                        'msg' => 'notFond'
+                    ]
+                );
+            }
+            if (!Gate::allows('view', $contact)) {
+                return redirect()->back();
+            }
+            return response()->json([
+                'status' => false,
+                'contact' => $contact
+            ]);
+        } catch (\Exception $e) {
+            return $e;
             return response()->json([
                 'status' => false,
                 'msg' => 'Not Found',
             ]);
-        } catch (NotAuthrizedException $e) {
-            return redirect()->back();
         }
     }
 
     public function update(ContactRequest $request, $id)
     {
         try {
-            $this->updateContact($request, $id);
-            return response()->json([
-                'status' => true,
+            $userId = Auth::id();
+            $contact = Contact::whereUserId($userId)->whereId($id)->first();
+            if(!$contact){
+                return response()->json([
+                        'status' => false,
+                        'msg' => 'notFond'
+                    ]
+                );
+            }
+            if (!Gate::allows('view', $contact)) {
+                return redirect()->back();
+            }
+
+            $contact->update([
+                'contact_string' => $request->contact,
+                'provider_id' => $request->providerId,
+                'user_id' => $userId
             ]);
-        } catch (NotFoundException $e) {
-            return response()->json([
-                'status' => false,
-                'msg' => 'Not Found',
-            ]);
-        } catch (NotAuthrizedException $e) {
-            return redirect()->back();
+            if ($contact) {
+                return response()->json([
+                    'status' => true,
+                    'msg' => 'added successfully',
+                ]);
+            }
+
+        } catch (\Exception $e) {
+            return $e;
         }
     }
 
     public function delete($id)
     {
         $contact = Contact::find($id);
+        if(!$contact){
+            return response()->json([
+                    'status' => false,
+                    'msg' => 'notFond'
+                ]
+            );
+        }
+        if (!Gate::allows('view', $contact)) {
+            return redirect()->back();
+        }
         $contact->delete();
         return response()->json([
             'status' => true,
